Fix documentation.
Turns out the built docs *do* ship, but in the `doc-xc/` directory, not
`doc/`. So find them there instead of building in doc/. This includes man
pages for GTM, so be sure to include those in the -gtm RPM. And also add the
postgres lang files to the GTM RPM, since the postgres binary is included in
it.